+++ mobile/android/android-components/components/feature/sitepermissions/src/main/java/mozilla/components/feature/sitepermissions/OnDiskSitePermissionsStorage.kt
@@ -89,7 +89,7 @@ class OnDiskSitePermissionsStorage(
         includeTemporary: Boolean,
         private: Boolean,
     ): SitePermissions? {
-        if (private) return SitePermissions(origin, savedAt = 0)
+        if (private) return null
         return database
             .sitePermissionsDao()
             .getSitePermissionsBy(origin)